Fix Paymob payment: cast items amount to integer
Paymob API requires items.amount to be a valid integer, not a float.
The multiplication of amount * 100 was producing a float which caused
the API to reject the request with "A valid integer is required."

+++ app/Services/Payment/DTOs/PaymentIntent.php
@@ -64,7 +64,7 @@ readonly class PaymentIntent
         if ($payment->payable) {
             $items[] = [
                 'name' => $payment->description ?? 'اشتراك',
-                'amount' => $payment->amount * 100, // Convert to cents
+                'amount' => (int) ($payment->amount * 100), // Convert to cents (must be integer for Paymob)
                 'quantity' => 1,
             ];
         }
